fixed useProjectCardAnimation issue
as it turns out, it was just a problem with the image used for the
second size, being severely oversized compared with the other two.
The originally intended animation can now be used freely and without
lags.

Also, divided usePortfolioBioFadeBlur for a smoother feel and the freedom
to apply hover states for the skill-icons on CSS